Add PagesController with home view
This commit creates a `PagesController`, two routes and one view.
The root points to `pages#home` where you can upload a plain text file
of the product list. This will then send a POST request to
`pages#upload`,which will then use the `ProductListFileHandler`
module to parse the contents and return a `Basket`. The view then uses
the `Basket` to display each item, the total (shelf prices - incl.
sales tax) and the total sales tax.